requirements.txt
fastapi[all]
uvicorn[standard]
pytesseract
opencv-python
opencv-python-headless
numpy
pillow
pandas
python-dotenv
pdf2image
tensorflow
pydantic
boto3